NP-1706: put online tests in separate file
